web/requirements.txt
asgiref==3.4.1
autopep8==1.5.7
dj-database-url==0.5.0
Django==3.2.19
dnspython==2.3.0
gunicorn==20.1.0
mysql-connector-python==8.0.32
mysqlclient==2.1.1
protobuf==3.20.3
pycodestyle==2.7.0
pymongo==4.3.3
python-decouple==3.4
pytz==2021.1
redis==4.5.5
sqlparse==0.4.4
toml==0.10.2
Unipath==1.1
whitenoise==5.3.0
